// api/neynar-api/models/UpdateUserReqBody.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SignerUUID } from './SignerUUID';
export type UpdateUserReqBody = {
    signer_uuid: SignerUUID;
    bio?: string;
    pfp_url?: string;
    url?: string;
    username?: string;
    display_name?: string;
};